+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 6.5.11 2024-??-??
  - 2024-09-18 Fixed exit code of migration script.
  - 2024-09-10 Fixed sorting of translated list entries in dropdown elements.
+ - 2024-09-09 Improved error logging in OAuth2 token overview. Fixed token error message reset.
  - 2024-08-27 Fixed console command Admin::Service::Add - does not work after ITSMCore is installed.
  - 2024-08-27 Added auto mapping for unknown actions to Znuny.Form.Input.FieldIDMapping.
  - 2024-08-22 Added '--single-transaction' option to scripts/backup.pl for MySQL/MariaDB. The --single-transaction flag will start a transaction before running.